// src/permission/seeder/domain/permission-categories.enum.ts
export enum PermissionCategoriesEnum {
    AUTH = "Authentication",
    ADJUSTMENT = "Adjustment",
    PERMISSION = "Permission",
    CANCER_STAGE = "Cancer Stage",
    CANCER_TYPE = "Cancer Type",
    DELIVERY_STATUS = "Delivery Status",
    HOSPITAL = "Hospital",
    INVOICE = "Invoice",
    LOCATIONS = "Locations",
    MANUFACTURER = "Manufacturer",
    ORDER = "Order",
    ORDER_STATUS = "Order Status",
    PATIENT = "Patient",
    PAYMENT = "Payment",
    PRESCRIPTION = "Prescription",
    PRICE= "Price",
    PROCUREMENT= "Procurement",
    PRODUCT = "Product",
    ROLE = "Role",
    STOCK = "Stock",
    STOCK_ISSUE = "Stock Issue",
    STOCK_RECEIPT = "Stock Receipt",
    USER = "User",
    CLIENT = "Client",
}